[backend/frontend] Refactor ingestion code + adapt UI to display logs
