+++ opencti-platform/opencti-front/src/private/components/data/IngestionTaxiis.tsx
@@ -56,7 +56,7 @@ const IngestionTaxii = () => {
   const dataColumns = {
     name: {
       label: 'Name',
-      width: '20%',
+      width: '25%',
       isSortable: true,
     },
     uri: {
@@ -66,12 +66,12 @@ const IngestionTaxii = () => {
     },
     ingestion_running: {
       label: 'Status',
-      width: '20%',
+      width: '10%',
       isSortable: false,
     },
     last_execution_date: {
       label: 'Last run',
-      width: '15%',
+      width: '20%',
       isSortable: false,
     },
     added_after_start: {
