+++ opencti-platform/opencti-front/src/private/components/data/ingestionJson/IngestionJsonLine.tsx
@@ -18,6 +18,7 @@ import type { Theme } from '../../../../components/Theme';
 import { INGESTION_SETINGESTIONS } from '../../../../utils/hooks/useGranted';
 import Security from '../../../../utils/Security';
 import { HandleAddFilter } from '../../../../utils/hooks/useLocalStorage';
+import IngestionTooltip from '../../../../components/IngestionTooltip';
 
 // Deprecated - https://mui.com/system/styles/basics/
 // Do not use it for new code.
@@ -58,6 +59,7 @@ const ingestionJsonLineFragment = graphql`
     connector_id
     ingestion_running
     last_execution_date
+    ingestionLogs
   }
 `;
 
@@ -67,7 +69,7 @@ export const IngestionJsonLineComponent: FunctionComponent<IngestionJsonLineProp
   paginationOptions,
 }) => {
   const classes = useStyles();
-  const { t_i18n, fldt } = useFormatter();
+  const { t_i18n, nsdt } = useFormatter();
   const data = useFragment(ingestionJsonLineFragment, node);
   return (
     <ListItem
@@ -115,7 +117,9 @@ export const IngestionJsonLineComponent: FunctionComponent<IngestionJsonLineProp
               className={classes.bodyItem}
               style={{ width: dataColumns.connector.width }}
             >
-              {fldt(data.last_execution_date) || '-'}
+              <IngestionTooltip logs={data.ingestionLogs}>
+                {nsdt(data.last_execution_date) || '-'}
+              </IngestionTooltip>
             </div>
             <div
               className={classes.bodyItem}
