+++ opencti-platform/opencti-front/src/schema/relay.schema.graphql
@@ -13119,6 +13119,7 @@ type IngestionRss implements InternalObject & BasicObject {
   current_state_date: DateTime
   last_execution_date: DateTime
   ingestion_running: Boolean
+  ingestionLogs: [String!]!
 }
 
 enum IngestionRssOrdering {
@@ -13321,6 +13322,7 @@ type IngestionCsv implements InternalObject & BasicObject {
   markings: [String!]
   toConfigurationExport: String!
   duplicateCsvMapper: CsvMapper!
+  ingestionLogs: [String!]!
 }
 
 enum IngestionCsvOrdering {
@@ -13429,6 +13431,7 @@ type IngestionJson implements InternalObject & BasicObject {
   ingestion_running: Boolean
   last_execution_date: DateTime
   markings: [String!]
+  ingestionLogs: [String!]!
 }
 
 enum IngestionJsonOrdering {
