+++ opencti-platform/opencti-graphql/src/modules/ingestion/ingestion-csv-resolver.ts
@@ -15,6 +15,7 @@ import {
   testCsvIngestionMapping,
 } from './ingestion-csv-domain';
 import { userAlreadyExists } from '../user/user-domain';
+import { getIngestionLogs } from './ingestion-taxii-collection-domain';
 import { loadCreator } from '../../database/members';
 
 const ingestionCsvResolvers: Resolvers = {
@@ -30,6 +31,7 @@ const ingestionCsvResolvers: Resolvers = {
     csvMapper: (ingestionCsv, _, context) => csvFeedGetCsvMapper(context, context.user, ingestionCsv),
     toConfigurationExport: (ingestionCsv, _, context) => csvFeedMapperExport(context, context.user, ingestionCsv),
     duplicateCsvMapper: (ingestionCsv, _, context) => csvFeedGetNewDuplicatedCsvMapper(context, context.user, ingestionCsv),
+    ingestionLogs: (ingestionCsv) => getIngestionLogs(ingestionCsv),
   },
   Mutation: {
     ingestionCsvTester: (_, { input }, context) => {
