+++ opencti-platform/opencti-graphql/src/modules/ingestion/ingestion-csv.graphql
@@ -25,6 +25,7 @@ type IngestionCsv implements InternalObject & BasicObject {
     markings: [String!]
     toConfigurationExport: String!
     duplicateCsvMapper: CsvMapper!
+    ingestionLogs: [String!]!
 }
 
 enum IngestionCsvOrdering {
