+++ opencti-platform/opencti-graphql/src/modules/ingestion/ingestion-json-resolver.ts
@@ -26,6 +26,7 @@ import {
   testJsonIngestionMapping,
 } from './ingestion-json-domain';
 import { connectorIdFromIngestId } from '../../domain/connector';
+import { getIngestionLogs } from './ingestion-taxii-collection-domain';
 import { loadCreator } from '../../database/members';
 
 const ingestionJsonResolvers: Resolvers = {
@@ -37,6 +38,7 @@ const ingestionJsonResolvers: Resolvers = {
     user: (ingestionJson, _, context) => loadCreator(context, context.user, ingestionJson.user_id),
     connector_id: (ingestionJson) => connectorIdFromIngestId(ingestionJson.id),
     jsonMapper: (ingestionJson, _, context) => findJsonMapperForIngestionById(context, context.user, ingestionJson.json_mapper_id),
+    ingestionLogs: (ingestionJson) => getIngestionLogs(ingestionJson),
   },
   Mutation: {
     ingestionJsonTester: (_, { input }, context) => {
