+++ opencti-platform/opencti-graphql/src/modules/ingestion/ingestion-json.graphql
@@ -42,6 +42,7 @@ type IngestionJson implements InternalObject & BasicObject {
     ingestion_running: Boolean
     last_execution_date: DateTime
     markings: [String!]
+    ingestionLogs: [String!]!
 }
 
 enum IngestionJsonOrdering {
