+++ opencti-platform/opencti-graphql/src/modules/ingestion/ingestion-rss-resolver.ts
@@ -4,6 +4,7 @@ import { storeLoadByIds } from '../../database/middleware-loader';
 import { ENTITY_TYPE_MARKING_DEFINITION } from '../../schema/stixMetaObject';
 import type { BasicStoreEntityMarkingDefinition } from '../../types/store';
 import { ENTITY_TYPE_IDENTITY } from '../../schema/general';
+import { getIngestionLogs } from './ingestion-taxii-collection-domain';
 import { loadCreator } from '../../database/members';
 
 const ingestionRssResolvers: Resolvers = {
@@ -16,6 +17,7 @@ const ingestionRssResolvers: Resolvers = {
     // eslint-disable-next-line max-len
     defaultMarkingDefinitions: (ingestionRss, _, context) => storeLoadByIds<BasicStoreEntityMarkingDefinition>(context, context.user, ingestionRss.object_marking_refs ?? [], ENTITY_TYPE_MARKING_DEFINITION),
     user: (ingestionRss, _, context) => loadCreator(context, context.user, ingestionRss.user_id),
+    ingestionLogs: (ingestionRss) => getIngestionLogs(ingestionRss),
   },
   Mutation: {
     ingestionRssAdd: (_, { input }, context) => {
