+++ opencti-platform/opencti-graphql/src/modules/ingestion/ingestion-rss.graphql
@@ -18,6 +18,7 @@ type IngestionRss implements InternalObject & BasicObject {
     current_state_date: DateTime
     last_execution_date: DateTime
     ingestion_running: Boolean
+    ingestionLogs: [String!]!
 }
 enum IngestionRssOrdering {
     name
