+++ opencti-platform/opencti-graphql/src/modules/ingestion/ingestion-taxii-collection-resolver.ts
@@ -1,4 +1,4 @@
-import { addIngestion, findTaxiiCollectionPaginated, findById, ingestionDelete, ingestionEditField, getIngestionLogs } from './ingestion-taxii-collection-domain';
+import { addIngestion, findTaxiiCollectionPaginated, findById, ingestionDelete, ingestionEditField } from './ingestion-taxii-collection-domain';
 import type { Resolvers } from '../../generated/graphql';
 import { getAuthorizedMembers } from '../../utils/authorizedMembers';
 import { loadCreator } from '../../database/members';
