+++ opencti-platform/opencti-graphql/src/modules/ingestion/ingestion-types.ts
@@ -8,6 +8,7 @@ import type { AuthorizedMember } from '../../utils/access';
 export const ENTITY_TYPE_INGESTION_RSS = 'IngestionRss';
 
 export interface BasicStoreEntityIngestionRss extends BasicStoreEntity {
+  kind: 'rss';
   name: string;
   description: string;
   scheduling_period: string;
@@ -46,6 +47,7 @@ export interface StixIngestionRss extends StixObject {
 export const ENTITY_TYPE_INGESTION_TAXII = 'IngestionTaxii';
 
 export interface BasicStoreEntityIngestionTaxii extends BasicStoreEntity {
+  kind: 'taxii';
   name: string;
   description: string;
   uri: string;
@@ -89,6 +91,7 @@ export interface StixIngestionTaxii extends StixObject {
 export const ENTITY_TYPE_INGESTION_CSV = 'IngestionCsv';
 
 export interface BasicStoreEntityIngestionCsv extends BasicStoreEntity {
+  kind: 'csv';
   current_state_hash: string;
   name: string;
   description: string;
@@ -130,6 +133,7 @@ export interface StixIngestionCsv extends StixObject {
 export const ENTITY_TYPE_INGESTION_JSON = 'IngestionJson';
 
 export interface BasicStoreEntityIngestionJson extends BasicStoreEntity {
+  kind: 'json';
   name: string;
   description: string;
   scheduling_period: string;
@@ -177,6 +181,7 @@ export interface StixIngestionJson extends StixObject {
 export const ENTITY_TYPE_INGESTION_TAXII_COLLECTION = 'IngestionTaxiiCollection';
 
 export interface BasicStoreEntityIngestionTaxiiCollection extends BasicStoreEntity {
+  kind: 'taxii-collection';
   name: string;
   description: string;
   user_id: string | undefined;
